extract util-functions to file
