+++ test/modules/map/components/olMap/handler/measure/OlMeasureUtils.test.js
@@ -1,0 +1,61 @@
+import { getGeometryLength, canShowAzimuthCircle } from '../../../../../../../src/modules/map/components/olMap/handler/measure/OlMeasureUtils';
+import { Point, LineString, Polygon, Circle, LinearRing } from 'ol/geom';
+describe('getGeometryLength', () => {
+	it('calculates length of LineString', () => {
+		const lineString = new LineString([[0, 0], [1, 0]]);
+		const length = getGeometryLength(lineString);
+
+		expect(length).toBe(1);
+	});
+
+	it('calculates length of LinearRing', () => {
+		const linearRing = new LinearRing([[0, 0], [1, 0], [1, 1], [0, 1], [0, 0]]);
+		const length = getGeometryLength(linearRing);
+
+		expect(length).toBe(4);
+	});
+
+	it('calculates length of Polygon', () => {		
+		const polygon = new Polygon([[[0, 0], [1, 0], [1, 1], [0, 1], [0, 0]]]);
+		const length = getGeometryLength(polygon);
+
+		expect(length).toBe(4);
+	});
+
+	
+	it('calculates not length of Circle', () => {
+		const circle = new Circle([0, 0], 1);
+		const length = getGeometryLength(circle);
+
+		expect(length).toBe(0);
+	});
+});
+
+describe('canShowAzimuthCircle', () => {
+	it('can show for a 2-point-line', () => {
+		const twoPointLineString = new LineString([[0, 0], [1, 0]]);
+
+		expect(canShowAzimuthCircle(twoPointLineString)).toBeTrue();
+	});
+
+	
+	it('can show for a pseudo-2-point-line', () => {
+		const threePointLineString = new LineString([[0, 0], [1, 0], [1, 0]]);
+
+		expect(canShowAzimuthCircle(threePointLineString)).toBeTrue();
+	});
+
+	
+	it('can NOT show for a point', () => {
+		const point = new Point([0, 0]);
+
+		expect(canShowAzimuthCircle(point)).toBeFalse();
+	});
+
+	
+	it('can NOT show for lineString', () => {
+		const threePointLineString = new LineString([[0, 0], [1, 0], [2, 1]]);
+
+		expect(canShowAzimuthCircle(threePointLineString)).toBeFalse();
+	});
+});
